회원정보 수정 탈퇴 추가

+++ src/service/MemberService.java
@@ -2,6 +2,8 @@ package service;
 
 import java.util.List;
 
+import javax.servlet.http.HttpSession;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
@@ -11,42 +13,52 @@ import vo.Member;
 @Component
 public class MemberService {
 	private MemberDao dao;
-	
+
 	@Autowired
 	public void setDao(MemberDao dao) {
 		this.dao = dao;
 	}
-	
-	public boolean insert(Member member){
-		if(dao.insertMember(member)>0){
+
+	public boolean insert(Member member) {
+		if (dao.insertMember(member) > 0) {
 			return true;
-		}else{
+		} else {
 			return false;
 		}
 	}
-	
-	public boolean update(Member member){
-		if(dao.updateMember(member)>0){
-			return true;
-		}else{
-			return false;
-		}		
-	}
-	
-	public boolean delete(String id){
-		if(dao.deleteMember(id)>0){
-			return true;
-		}else{
-			return false;
+
+	public boolean update(Member member, HttpSession session) {
+		Member original = dao.selectMember(member.getId());
+		String password = (String) session.getAttribute("password");
+		if (original.getPassword().equals(password)) {
+			if (dao.updateMember(member) > 0) {
+				return true;
+			} else {
+				return false;
+			}
 		}
+		return false;
 	}
-	
-	public Member select(String id){
+
+	public boolean delete(String id, HttpSession session) {
+		Member original = dao.selectMember(id);
+		String password = (String) session.getAttribute("password");
+		if (original.getPassword().equals(password)) {
+			if (dao.deleteMember(id) > 0) {
+				return true;
+			} else {
+				return false;
+			}
+		}
+		return false;
+	}
+
+	public Member select(String id) {
 		Member member = dao.selectMember(id);
 		return member;
 	}
-	
-	public List<Member> selectList(){
+
+	public List<Member> selectList() {
 		List<Member> result = dao.selectListMember();
 		return result;
 	}
